+++ CHANGELOG.md
@@ -78,6 +78,8 @@
   ([GH-1618](https://github.com/NVIDIA/warp/issues/1618)).
 - Add an example of a distributed Jacobi solver that uses NCCL (via `nccl4py`) for nearest-neighbor halo exchange in
   `warp/examples/distributed/example_jacobi_nccl.py` ([GH-1576](https://github.com/NVIDIA/warp/issues/1576)).
+- Add an example of a distributed Jacobi solver that uses NVSHMEM (via `nvshmem4py`) for nearest-neighbor halo exchange
+  in `warp/examples/distributed/example_jacobi_nvshmem.py` ([GH-1582](https://github.com/NVIDIA/warp/issues/1582)).
 
 ## [1.15.0] - 2026-07-07